afxdp/cos/cross_binding: relocate inline tests to cross_binding_tests.rs
Same playbook as #1080-#1100. Extract 589-LOC inline test block from
cos/cross_binding.rs to sibling cross_binding_tests.rs.

After: cross_binding.rs drops from 866 → 279 LOC (~3× reduction).

cargo build --release: clean. cargo test --release: 878 passed.
